requirements.txt
flask
flask-restful
flask-migrate
flask-sqlalchemy
flask-marshmallow
marshmallow-sqlalchemy
python-dotenv
passlib
tox
pytest
pytest-runner
pytest-flask
pytest-factoryboy
factory_boy
